Make sure ES6 registerables (like Stores, StateSources and ActionCreators) get their `id` to default to their class name.
[ref](https://gitter.im/jhollingworth/marty?at=5512b9e2a61b930e5dbca11a)
[manual test to showcase the issue](https://gist.github.com/dariocravero/c4442ad42412a43ec1ff)
