+++ lib/registry.js
@@ -67,7 +67,8 @@ class Registry {
     }
 
     clazz.id = id;
-    clazz.type = type;
+    defaultInstance.id = defaultInstance.id || id;
+    defaultInstance.type = clazz.type = type;
 
     this.types[type][id] = clazz;
